+++ src/scene/camera.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 #include "camera.h"
 
-Camera::Camera(Point center, Point target, Vector up, float vertical_fov,
+Camera::Camera(Point center, Point target, Vector up, double vertical_fov,
                uint32_t pixel_height, uint32_t pixel_width) : center { center },
                                                               target { target },
                                                               up { up },
@@ -11,7 +11,7 @@ Camera::Camera(Point center, Point target, Vector up, float vertical_fov,
                                                               pixel_width { pixel_width }
 {
     assert(pixel_height != 0 && pixel_width != 0);
-    aspect_ratio = static_cast<float>(pixel_width) / static_cast<float>(pixel_height);
+    aspect_ratio = static_cast<double>(pixel_width) / static_cast<double>(pixel_height);
 
     sensor_height = 2.0f * std::tan(vertical_fov / 2.0f);
     sensor_width = aspect_ratio * sensor_height;
@@ -30,8 +30,8 @@ Camera::Camera(Point center, Point target, Vector up, float vertical_fov,
 
 Ray Camera::cast_ray(const uint32_t& px, const uint32_t& py) const
 {
-    float sx = (px * sensor_width) / (pixel_width - 1);
-    float sy = (py * sensor_height) / (pixel_height - 1);
+    double sx = (px * sensor_width) / (pixel_width - 1);
+    double sy = (py * sensor_height) / (pixel_height - 1);
 
     Point pixel = lower_left_pixel + sx * u + sy * v;
     Vector direction = (pixel - center).normalized();
